Some code refactor and moving of files. All tests are passing.
