+++ lib/bitcoin/interfaces/ISlidingWindowQuantileStore.ts
