+++ lib/bitcoin/models/ProofOfFeeConfig.ts
