+++ lib/common/models/TransactionFeeModel.ts
@@ -1,6 +1,6 @@
 /**
  * Defines Sidetree minimum proof-of-fee
  */
-export default interface FeeModel {
+export default interface TransactionFeeModel {
   normalizedTransactionFee: number;
 }
